+++ packages/react-ui/src/widgets/TableWidgetUI/TableWidgetUI.js
@@ -20,7 +20,8 @@ import useSkeleton from '../useSkeleton';
 
 const TableHeadCellLabel = styled(TableSortLabel)(({ theme }) => ({
   ...theme.typography.caption,
-  color: theme.palette.text.secondary
+  color: theme.palette.text.secondary,
+  whiteSpace: 'nowrap'
 }));
 
 const TableRowStyled = styled(TableRow)(({ theme }) => ({
@@ -64,6 +65,8 @@ function TableWidgetUI({
   rowsPerPageOptions,
   onSetRowsPerPage,
   onRowClick,
+  onRowMouseEnter,
+  onRowMouseLeave,
   height,
   dense,
   isLoading,
@@ -126,7 +129,13 @@ function TableWidgetUI({
             sortDirection={sortDirection}
             onSort={handleSort}
           />
-          <TableBodyComponent columns={columns} rows={rows} onRowClick={onRowClick} />
+          <TableBodyComponent
+            columns={columns}
+            rows={rows}
+            onRowMouseEnter={onRowMouseEnter}
+            onRowMouseLeave={onRowMouseLeave}
+            onRowClick={onRowClick}
+          />
         </Table>
       </TableContainer>
       {pagination && (
@@ -182,7 +191,13 @@ function TableHeaderComponent({ columns, sorting, sortBy, sortDirection, onSort 
   );
 }
 
-function TableBodyComponent({ columns, rows, onRowClick }) {
+function TableBodyComponent({
+  columns,
+  rows,
+  onRowMouseEnter,
+  onRowMouseLeave,
+  onRowClick
+}) {
   return (
     <TableBody>
       {rows.map((row, i) => {
@@ -192,13 +207,17 @@ function TableBodyComponent({ columns, rows, onRowClick }) {
           <TableRowStyled
             key={rowKey}
             hover={!!onRowClick}
+            onMouseEnter={() => onRowMouseEnter && onRowMouseEnter(row)}
+            onMouseLeave={() => onRowMouseLeave && onRowMouseLeave(row)}
             onClick={() => onRowClick && onRowClick(row)}
           >
-            {columns.map(({ field, headerName, align, component }) => {
+            {columns.map(({ field, headerName, align, component, formatter }) => {
               let cellValue = Object.entries(row).find(([key]) => {
                 return key.toUpperCase() === field.toUpperCase();
               })?.[1];
-              if (typeof cellValue === 'bigint') {
+              if (formatter) {
+                cellValue = formatter(cellValue);
+              } else if (typeof cellValue === 'bigint') {
                 cellValue = cellValue.toString(); // otherwise TableCell will fail for displaying it
               } else if (Array.isArray(cellValue)) {
                 cellValue = `[${cellValue
@@ -251,6 +270,8 @@ TableWidgetUI.propTypes = {
   rowsPerPageOptions: PropTypes.array,
   onSetRowsPerPage: PropTypes.func,
   onRowClick: PropTypes.func,
+  onRowMouseEnter: PropTypes.func,
+  onRowMouseLeave: PropTypes.func,
   height: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
   dense: PropTypes.bool,
   isLoading: PropTypes.bool,