+++ CHANGELOG.md
@@ -3,12 +3,23 @@
 ## Not released
 
 - TileFeatures returns features with geometry if asked [#911](https://github.com/CartoDB/carto-react/pull/911)
+- Support for `onRowMouseEnter` and `onRowMouseLeave` handlers for Table Widget [#907](https://github.com/CartoDB/carto-react/pull/907)
+
+## 3.0.0
+
+### 3.0.0-alpha.20 (2024-09-20)
+
+- `NoWrap` for table header [#908](https://github.com/CartoDB/carto-react/pull/908)
+
+### 3.0.0-alpha.19 (2024-09-19)
+
 - Fix HistogramWidget with one non-zero bucket [#901](https://github.com/CartoDB/carto-react/pull/901)
 - Spatial Index Sources use remote widgets calculation [#898](https://github.com/CartoDB/carto-react/pull/898)
 - Support for `hiddenColumnFields` parameter and `onRowClick` handler for Table Widget [#900](https://github.com/CartoDB/carto-react/pull/900)
-- Support for `searchText` and `searchColumn` for Table Widget [#900](https://github.com/CartoDB/carto-react/pull/902)
-
-## 3.0.0
+- Support for `searchText` and `searchColumn` for remote calculations for Table Widget [#902](https://github.com/CartoDB/carto-react/pull/902)
+- Support `searchText` and `searchColumn` for local calculations (Tilesets) for Table Widget [#903](https://github.com/CartoDB/carto-react/pull/903)
+- Support for columns `formatter` function for Table Widget [#904](https://github.com/CartoDB/carto-react/pull/904)
+- isRemoteCalculationSupported: revert remote calculation for dynamic spatial index sources [#905](https://github.com/CartoDB/carto-react/pull/905)
 
 ### 3.0.0-alpha.17 (2024-07-29)